before:
```
name ............................... min time      avg time    std dv   runs
findAll ............................ 0.413 ms      0.777 ms    ±0.276  x1000
```

after
```
name ............................... min time      avg time    std dv   runs
findAll ............................ 0.081 ms      0.442 ms    ±0.273  x1000
```